add events of course method

+++ internal/services/course/service/service.go
@@ -5,15 +5,18 @@ import (
 	"dussh/internal/cache/redis"
 	"dussh/internal/domain/models"
 	coursev1 "dussh/internal/services/course/api/v1"
+	"errors"
 	"go.uber.org/zap"
 )
 
 type Repository interface {
 	GetCourse(ctx context.Context, courseID int64) (*models.Course, error)
 	SaveCourse(ctx context.Context, crs *models.Course) (int64, error)
+	SaveEvents(ctx context.Context, courseID int64, events []*models.Event) error
 	UpdateCourse(ctx context.Context, id int64, crs *models.Course) error
 	DeleteCourse(ctx context.Context, id int64) error
 	DeleteEvent(ctx context.Context, courseID, eventID int64) error
+	CheckCountEvents(ctx context.Context, courseID int64) (int, error)
 }
 
 func NewCourseService(
@@ -25,6 +28,10 @@ func NewCourseService(
 		log:  log.Named("course.service"),
 	}
 }
+
+var (
+	ErrMustBeAtLeastOneEvent = errors.New("there must be at least one event")
+)
 
 type courseService struct {
 	repo  Repository
@@ -41,6 +48,10 @@ func (c *courseService) Create(ctx context.Context, crs *models.Course) (int64, 
 	return c.repo.SaveCourse(ctx, crs)
 }
 
+func (c *courseService) AddEvents(ctx context.Context, courseID int64, events []*models.Event) error {
+	return c.repo.SaveEvents(ctx, courseID, events)
+}
+
 func (c *courseService) Update(ctx context.Context, id int64, crs *models.Course) error {
 	return c.repo.UpdateCourse(ctx, id, crs)
 }
@@ -50,5 +61,14 @@ func (c *courseService) Delete(ctx context.Context, id int64) error {
 }
 
 func (c *courseService) DeleteEvent(ctx context.Context, courseID, eventID int64) error {
+	count, err := c.repo.CheckCountEvents(ctx, courseID)
+	if err != nil {
+		return err
+	}
+
+	if count <= 1 {
+		return ErrMustBeAtLeastOneEvent
+	}
+
 	return c.repo.DeleteEvent(ctx, courseID, eventID)
 }
